+++ src/main/java/com/github/evillootlye/caves/PlayerAttackedEvent.java
@@ -1,0 +1,41 @@
+package com.github.evillootlye.caves;
+
+import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Player;
+import org.bukkit.event.Cancellable;
+import org.bukkit.event.HandlerList;
+import org.bukkit.event.player.PlayerEvent;
+
+public class PlayerAttackedEvent extends PlayerEvent implements Cancellable {
+    private static final HandlerList HANDLERS = new HandlerList();
+    private final LivingEntity attacker;
+    private boolean cancelled;
+
+    public PlayerAttackedEvent(Player player, LivingEntity attacker) {
+        super(player);
+        this.attacker = attacker;
+    }
+
+    public LivingEntity getAttacker() {
+        return attacker;
+    }
+
+    @Override
+    public boolean isCancelled() {
+        return cancelled;
+    }
+
+    @Override
+    public void setCancelled(boolean cancelled) {
+        this.cancelled = cancelled;
+    }
+
+    @Override
+    public HandlerList getHandlers() {
+        return HANDLERS;
+    }
+
+    public static HandlerList getHandlerList() {
+        return HANDLERS;
+    }
+}
