+++ src/main/java/com/github/evillootlye/caves/configuration/Configuration.java
@@ -1,0 +1,4 @@
+package com.github.evillootlye.caves.configuration;
+
+public class Configuration {
+}
