+++ src/main/java/com/github/evillootlye/caves/generator/StructureFiles.java
@@ -1,0 +1,54 @@
+package com.github.evillootlye.caves.generator;
+
+public class StructureFiles {
+
+    public static int[][][] zerozerotwo = { { {1, 1, 1, 1, 1, 1, 1, 1, 1}, {2, 2, 3, 3, 3, 2, 2, 2, 2},{2, 2, 4, 4, 4, 2, 2, 2, 2},    {2, 2, 4, 4, 4, 2, 2, 2, 2},{2, 2, 4, 4, 4, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2},{1, 1, 1, 1, 1, 1, 1, 1, 1},{1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 2, 2, 2, 2, 2, 2, 2, 1}, {2, 3, 3, 3, 3, 3, 3, 3, 2},{2, 10, 11, 0, 0, 9, 9, 9, 2}, {2, 9, 11, 0, 0, 0, 18, 17, 2},{2, 17, 11, 0, 0, 14, 14, 14, 2},{2, 0, 11, 0, 0, 0, 0, 0, 2},{1, 2, 2, 2, 2, 2, 2, 2, 1},{1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 2, 2, 2, 2, 2, 2, 2, 1}, {2, 3, 3, 3, 3, 3, 3, 3, 3},{2, 0, 0, 19, 19, 19, 0, 0, 4},    {2, 0, 0, 0, 0, 0, 0, 0, 4},{2, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 2},{1, 2, 0, 0, 0, 14, 0, 2, 1},{1, 1, 5, 5, 5, 5, 5, 1, 1} },
+            {   {1, 2, 2, 2, 2, 2, 2, 2, 1}, {3, 3, 3, 3, 3, 3, 3, 3, 2},{4, 0, 0, 19, 0, 19, 0, 15, 2},    {4, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 2},{1, 2, 0, 0, 0, 14, 0, 2, 1},{1, 1, 5, 5, 5, 5, 5, 1, 1} },
+            {   {1, 2, 2, 2, 2, 2, 2, 2, 1}, {2, 3, 3, 3, 3, 3, 3, 3, 2},{2, 0, 0, 19, 0, 19, 0, 16, 2},    {2, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 14, 2},{2, 11, 11, 11, 11, 11, 11, 11, 2},{1, 2, 0, 0, 0, 14, 0, 2, 1},{1, 1, 5, 5, 5, 5, 5, 1, 1} },
+            {   {1, 2, 2, 2, 2, 2, 2, 2, 1}, {2, 3, 3, 3, 3, 3, 3, 3, 2},{2, 12, 0, 19, 19, 19, 0, 15, 2},   {2, 12, 0, 0, 0, 0, 0, 0, 2},{2, 12, 0, 0, 0, 0, 0, 14, 2},{2, 0, 0, 0, 0, 0, 0, 0, 2},{1, 2, 0, 0, 0, 14, 0, 2, 1},{1, 1, 5, 5, 5, 5, 5, 1, 1} },
+            {   {1, 2, 2, 2, 2, 2, 2, 2, 1}, {2, 3, 3, 3, 3, 3, 3, 3, 2},{2, 10, 0, 0, 13, 13, 0, 14, 2},{2, 10, 0, 0, 19, 19, 0, 17, 2},{2, 10, 0, 0, 14, 14, 0, 14, 2},{2, 0, 0, 0, 0, 0, 0, 0, 2},{1, 2, 2, 2, 2, 2, 2, 2, 1},{1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1}, {2, 2, 3, 2, 2, 2, 2, 2, 2},{2, 2, 0, 2, 2, 2, 2, 2, 2},    {2, 2, 0, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2},{1, 1, 1, 1, 1, 1, 1, 1, 1},{1, 1, 1, 1, 1, 1, 1, 1, 1} } };
+
+    public static int[][][] zerozerotwo2 = { { {2, 2, 2, 2, 2, 2, 2, 2, 2, 2}, {2, 2, 2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 6, 8, 8, 6, 2},{2, 2, 2, 2, 2, 6, 8, 8, 6, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2, 2, 2}, {2, 3, 3, 3, 3, 3, 3, 3, 3, 2},{2, 0, 0, 0, 18, 20, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 7, 7, 7, 7, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2, 2, 2}, {2, 3, 3, 3, 3, 3, 3, 3, 3, 2},{2, 0, 20, 0, 0, 0, 0, 18, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2, 2, 2}, {3, 3, 3, 3, 3, 3, 3, 3, 3, 2},{2, 18, 0, 0, 0, 0, 0, 0, 20, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2, 2, 2}, {2, 3, 3, 3, 3, 3, 3, 3, 3, 2},{2, 0, 0, 0, 20, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2, 2, 2}, {2, 3, 3, 3, 3, 3, 3, 3, 3, 2},{2, 20, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2, 2, 2}, {2, 3, 3, 3, 3, 3, 3, 3, 3, 2},{2, 0, 0, 0, 0, 18, 0, 20, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 0, 0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2, 2, 2} } };
+
+    public static int[][][] zerozerotwo3 = { { {1, 1, 1, 1, 1, 1, 1, 1}, {2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 1}, {3, 3, 3, 3, 3, 3, 3, 2},{0, 0, 0, 20, 0, 18, 20, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 1}, {3, 3, 3, 3, 3, 3, 3, 2},{0, 20, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 1}, {3, 3, 3, 3, 3, 3, 3, 2},{0, 18, 0, 20, 0, 20, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 1}, {3, 3, 3, 3, 3, 3, 3, 2},{0, 0, 18, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 1}, {3, 3, 3, 3, 3, 3, 3, 2},{20, 0, 20, 0, 0, 0, 20, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 1}, {3, 3, 3, 3, 3, 3, 3, 2},{0, 0, 0, 0, 20, 0, 18, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{0, 0, 0, 0, 0, 0, 0, 2},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1}, {2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2} } };
+
+    public static int[][][] zerozerotwo4 = { { {1, 1, 1, 1, 1, 1, 1, 1}, {2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2}, {2, 3, 3, 3, 3, 3, 3, 3},{2, 0, 0, 20, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2}, {2, 3, 3, 3, 3, 3, 3, 3},{2, 18, 0, 0, 18, 20, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2}, {2, 3, 3, 3, 3, 3, 3, 3},{2, 20, 0, 0, 0, 0, 0, 18},{2, 0, 0, 0, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2}, {2, 3, 3, 3, 3, 3, 3, 3},{2, 0, 20, 0, 0, 0, 0, 20},{2, 0, 0, 0, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {2, 2, 2, 2, 2, 2, 2, 2}, {2, 3, 3, 3, 3, 3, 3, 3},{2, 0, 18, 0, 20, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 0, 0, 0, 0, 0, 0, 0},{2, 2, 2, 2, 2, 2, 2, 2} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1}, {2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2},{2, 2, 2, 2, 2, 2, 2, 2} } };
+
+    public static int[][][] zerozerothree = { { {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} },
+            {   {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}, {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1} } };
+
+
+}
