+++ src/main/java/com/github/evillootlye/caves/mobs/DefaultMobs.java
@@ -1,0 +1,14 @@
+package com.github.evillootlye.caves.mobs;
+
+enum DefaultMobs {
+    ALPHA_SPIDER(new AlphaSpider()), HEXED_ARMOR(new HexedArmor());
+    private final CustomMob mob;
+
+    DefaultMobs(CustomMob mob) {
+        this.mob = mob;
+    }
+
+    public CustomMob getMob() {
+        return mob;
+    }
+}
