+++ src/main/java/com/github/evillootlye/caves/utils/Rnd.java
@@ -1,0 +1,34 @@
+package com.github.evillootlye.caves.utils;
+
+import java.util.concurrent.ThreadLocalRandom;
+
+public class Rnd {
+
+    public static boolean nextBoolean() {
+        return ThreadLocalRandom.current().nextBoolean();
+    }
+
+    public static int nextInt() {
+        return ThreadLocalRandom.current().nextInt();
+    }
+
+    public static int nextInt(int max) {
+        return ThreadLocalRandom.current().nextInt(max);
+    }
+
+    public static int nextInt(int min, int max) {
+        return ThreadLocalRandom.current().nextInt(min, max);
+    }
+
+    public static double nextDouble() {
+        return ThreadLocalRandom.current().nextDouble();
+    }
+
+    public static double nextDouble(double max) {
+        return ThreadLocalRandom.current().nextDouble(max);
+    }
+
+    public static double nextDouble(double min, double max) {
+        return ThreadLocalRandom.current().nextDouble(min, max);
+    }
+}
