+++ src/main/java/com/github/evillootlye/caves/utils/Utils.java
@@ -1,0 +1,8 @@
+package com.github.evillootlye.caves.utils;
+
+public class Utils {
+    @SafeVarargs
+    public static <T> T[] array(T... t) {
+        return t;
+    }
+}
